Split common imported files into emissions_data_lib from emissions_data
