+++ app/data-loader/src/models/emissionsFactor.ts
@@ -1,4 +1,4 @@
-import { EmissionsFactorInterface } from '@blockchain-carbon-accounting/emissions_data_chaincode/src/lib/emissionsFactor';
+import type { EmissionsFactorInterface } from '@blockchain-carbon-accounting/emissions_data_lib/src/emissionsFactor';
 import {
   Column, Entity,
   PrimaryGeneratedColumn
