+++ app/data-loader/src/repositories/emissionsFactor.repo.ts
@@ -1,9 +1,10 @@
 import { EmissionFactorDbInterface } from "@blockchain-carbon-accounting/data-common/db"
 import { ActivityInterface, getYearFromDate } from "@blockchain-carbon-accounting/data-common/utils"
-import { CO2EmissionFactorInterface, getUomFactor } from "@blockchain-carbon-accounting/emissions_data_chaincode/src/lib/emissions-calc"
-import { EmissionsFactorInterface, EMISSIONS_FACTOR_CLASS_IDENTIFER } from "@blockchain-carbon-accounting/emissions_data_chaincode/src/lib/emissionsFactor"
-import { UtilityLookupItemInterface } from "@blockchain-carbon-accounting/emissions_data_chaincode/src/lib/utilityLookupItem"
-import { ErrInvalidFactorForActivity } from "@blockchain-carbon-accounting/emissions_data_chaincode/src/util/const"
+import { CO2EmissionFactorInterface, getUomFactor } from "@blockchain-carbon-accounting/emissions_data_lib/src/emissions-calc"
+import { EMISSIONS_FACTOR_CLASS_IDENTIFER } from "@blockchain-carbon-accounting/emissions_data_lib/src/emissionsFactor"
+import type { EmissionsFactorInterface } from "@blockchain-carbon-accounting/emissions_data_lib/src/emissionsFactor"
+import type { UtilityLookupItemInterface } from "@blockchain-carbon-accounting/emissions_data_lib/src/utilityLookupItem"
+import { ErrInvalidFactorForActivity } from "@blockchain-carbon-accounting/emissions_data_lib/src/const"
 import { Between, DataSource, FindOptionsWhere, ILike, LessThanOrEqual, MoreThanOrEqual, SelectQueryBuilder } from "typeorm"
 import { EmissionsFactor } from "../models/emissionsFactor"
 import { UtilityLookupItem } from "../models/utilityLookupItem"
