+++ app/frontend/react-app/package.json
@@ -71,6 +71,7 @@
     "@types/react-copy-to-clipboard": "^5.0.2",
     "@types/react-dom": "^18.0.5",
     "@blockchain-carbon-accounting/api-server": "*",
+    "@blockchain-carbon-accounting/emissions_data_lib": "*",
     "crypto-browserify": "^3.12.0",
     "https-browserify": "^1.0.0",
     "node-libs-browser": "^2.2.1",
