+++ emissions-data/chaincode/emissionscontract/typescript/package.json
@@ -32,6 +32,7 @@
     "typescript": "^4.7.2"
   },
   "dependencies": {
+    "@blockchain-carbon-accounting/emissions_data_lib": "*",
     "crypto-js": "^4.1.1",
     "fabric-shim": "2.2.3"
   }
