+++ emissions-data/chaincode/emissionscontract/typescript/src/lib/emissionsFactor.ts
@@ -9,43 +9,13 @@
 
 /* tslint:disable:max-classes-per-file */
 import { ChaincodeStub } from 'fabric-shim';
-import { ErrStateNotFound } from '../util/const';
+import { ErrStateNotFound } from '@blockchain-carbon-accounting/emissions_data_lib/src/const';
 import { State } from '../util/state';
 import { QueryResult, WorldState } from '../util/worldstate';
-import { getYearFromDate } from './emissions-calc';
-import { UtilityLookupItemInterface } from './utilityLookupItem';
-
-export const EMISSIONS_FACTOR_CLASS_IDENTIFER =
-    'org.hyperledger.blockchain-carbon-accounting.emissionsfactoritem';
-
-export interface EmissionsFactorInterface {
-    class: string;
-    key?: string;
-    uuid: string;
-    type: string;
-    scope?: string;
-    level_1: string;
-    level_2: string;
-    level_3: string;
-    level_4?: string;
-    text?: string;
-    year?: string;
-    from_year?: string;
-    thru_year?: string;
-    country?: string;
-    division_type?: string;
-    division_id?: string;
-    division_name?: string;
-    activity_uom?: string;
-    net_generation?: string;
-    net_generation_uom?: string;
-    co2_equivalent_emissions?: string;
-    co2_equivalent_emissions_uom?: string;
-    source?: string;
-    non_renewables?: string;
-    renewables?: string;
-    percent_of_renewables?: string;
-}
+import { getYearFromDate } from '@blockchain-carbon-accounting/emissions_data_lib/src/emissions-calc';
+import type { UtilityLookupItemInterface } from '@blockchain-carbon-accounting/emissions_data_lib/src/utilityLookupItem';
+import type { EmissionsFactorInterface } from '@blockchain-carbon-accounting/emissions_data_lib/src/emissionsFactor';
+import { EMISSIONS_FACTOR_CLASS_IDENTIFER } from '@blockchain-carbon-accounting/emissions_data_lib/src/emissionsFactor';
 
 export class EmissionsFactor extends State {
     factor: EmissionsFactorInterface;
