+++ lib/emissions_data/src/const.ts
@@ -29,3 +29,4 @@ export const REGIONS = {
     // Countries
     USA: 'usa',
 };
+
