+++ lib/supply-chain/package.json
@@ -14,13 +14,13 @@
     "@types/node": "^17.0.40",
     "@typescript-eslint/eslint-plugin": "^5.27.0",
     "@typescript-eslint/parser": "^5.27.0",
+    "@blockchain-carbon-accounting/emissions_data_lib": "*",
     "eslint": "^8.17.0",
     "typescript": "^4.7.3"
   },
   "dependencies": {
     "@blockchain-carbon-accounting/data-postgres": "*",
     "@blockchain-carbon-accounting/emissions_data": "*",
-    "@blockchain-carbon-accounting/emissions_data_chaincode": "*",
     "@googlemaps/google-maps-services-js": "^3.3.14",
     "@metamask/eth-sig-util": "^4.0.1",
     "axios": "^0.27.2",
