+++ lib/supply-chain/src/common-types.ts
@@ -1,4 +1,4 @@
-import { EmissionsFactorInterface } from "@blockchain-carbon-accounting/emissions_data_chaincode/src/lib/emissionsFactor";
+import type { EmissionsFactorInterface } from "@blockchain-carbon-accounting/emissions_data_lib/src/emissionsFactor";
 import { LatLngLiteral } from "@googlemaps/google-maps-services-js";
 import { UpsResponse } from './ups-types';
