+++ net-emissions-token-network/test/emissions-factor-unit-tests.ts
@@ -1,6 +1,6 @@
 import { PostgresDBService } from '@blockchain-carbon-accounting/data-postgres/src/postgresDbService';
-import type { EmissionsFactorInterface } from '@blockchain-carbon-accounting/emissions_data_chaincode/src/lib/emissionsFactor';
-import type { UtilityLookupItemInterface } from "@blockchain-carbon-accounting/emissions_data_chaincode/src/lib/utilityLookupItem";
+import type { EmissionsFactorInterface } from '@blockchain-carbon-accounting/emissions_data_lib/src/emissionsFactor';
+import type { UtilityLookupItemInterface } from "@blockchain-carbon-accounting/emissions_data_lib/src/utilityLookupItem";
 import { Activity } from '@blockchain-carbon-accounting/supply-chain-lib/src/common-types';
 import { process_activity } from '@blockchain-carbon-accounting/supply-chain-lib/src/emissions-utils';
 import { expect } from 'chai';
